+++ src/app/layout.js
@@ -1,9 +1,9 @@
 import Footer from "@/components/Footer/Footer";
-import Header from "@/components/Header/Header";
+import Header from "@/components/Header/HeaderComponent/Header";
 import "./globals.scss";
 import { AppRouterCacheProvider } from "@mui/material-nextjs/v13-appRouter";
 import { ThemeProvider } from "@mui/material/styles";
-import theme from "./fontMUI";
+import theme from "./theme";
 import { Roboto } from "next/font/google";
 
 export const metadata = {
