+++ src/components/Header/HeaderTabsComponent/HeaderTabsUpper.config.js
